Get numpy include dir path from within build_ext class
Currently numpy is imported before it is installed as a dependecy.
As a workaround, append results of `numpy.get_include` method to
the `build_ext.include_dirs` list in the customized `build_ext` class.

To obtain numpy in the first place it is necesssary to include it in the
`setup_requires` section